refactor: Minor code and dependency injection cleanup
