+++ data/src/commonMain/kotlin/com/moneyplusplus/data/datasource/remote/client/createHttpClient.kt
@@ -1,14 +1,13 @@
-package com.moneyplusplus.data.util.network
+package com.moneyplusplus.data.datasource.remote.client
 
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.HttpClientEngine
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.plugins.logging.Logging
-import io.ktor.client.plugins.logging.LogLevel
 import io.ktor.client.plugins.HttpTimeout
-import io.ktor.http.ContentType.Application.Json
-import kotlinx.serialization.json.Json
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.logging.LogLevel
+import io.ktor.client.plugins.logging.Logging
 import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
 
 fun createHttpClient(
     engine: HttpClientEngine
