+++ data/src/commonMain/kotlin/com/moneyplusplus/data/di/dataModule.kt
@@ -3,5 +3,5 @@ package com.moneyplusplus.data.di
 import org.koin.dsl.module
 
 val dataModule = module {
-    includes(repositoryModule, platformModule())
+    includes(repositoryModule, platformModule)
 }
